Test 2 data disks

+++ azuredeploy.sh
@@ -11,8 +11,8 @@ whoami >> /tmp/azuredeploy.log.$$ 2>&1
 echo $@ >> /tmp/azuredeploy.log.$$ 2>&1
 
 # Usage
-if [ "$#" -ne 9 ]; then
-  echo "Usage: $0 MASTER_NAME MASTER_IP WORKER_NAME WORKER_IP_BASE WORKER_IP_START NUM_OF_VM ADMIN_USERNAME ADMIN_PASSWORD TEMPLATE_BASE" >> /tmp/azuredeploy.log.$$
+if [ "$#" -ne 10 ]; then
+  echo "Usage: $0 MASTER_NAME MASTER_IP WORKER_NAME WORKER_IP_BASE WORKER_IP_START NUM_OF_VM ADMIN_USERNAME ADMIN_PASSWORD NUM_OF_DATA_DISKS TEMPLATE_BASE" >> /tmp/azuredeploy.log.$$
   exit 1
 fi
 
@@ -20,20 +20,32 @@ fi
 ###################################
 
 # Parameters
-MASTER_NAME=$1
-MASTER_IP=$2
-WORKER_NAME=$3
-WORKER_IP_BASE=$4
-WORKER_IP_START=$5
-NUM_OF_VM=$6
-ADMIN_USERNAME=$7
-ADMIN_PASSWORD=$8
-TEMPLATE_BASE=$9
+MASTER_NAME=${1}
+MASTER_IP=${2}
+WORKER_NAME=${3}
+WORKER_IP_BASE=${4}
+WORKER_IP_START=${5}
+NUM_OF_VM=${6}
+ADMIN_USERNAME=${7}
+ADMIN_PASSWORD=${8}
+NUM_OF_DATA_DISKS=${9}
+TEMPLATE_BASE=${10}
 
 # Format and mount the data disk on master, install NFS, and create a cluster wide NFS share
-sudo sh -c "mkfs -t ext4 /dev/sdc" >> /tmp/azuredeploy.log.$$ 2>&1
-sudo sh -c "mkdir /data" >> /tmp/azuredeploy.log.$$ 2>&1
-echo "/dev/sdc /data ext4  defaults,discard 0 0" | sudo tee -a /etc/fstab >> /tmp/azuredeploy.log.$$ 2>&1
+if [$NUM_OF_DATA_DISKS eq "1" ]; then
+  sudo sh -c "mkfs -t ext4 /dev/sdc" >> /tmp/azuredeploy.log.$$ 2>&1
+  sudo sh -c "mkdir /data" >> /tmp/azuredeploy.log.$$ 2>&1
+  echo "/dev/sdc /data ext4  defaults,discard 0 0" | sudo tee -a /etc/fstab >> /tmp/azuredeploy.log.$$ 2>&1
+fi
+
+if [$NUM_OF_DATA_DISKS eq "2"]; then
+  sudo sh -c "mkfs -t ext4 /dev/sdc" >> /tmp/azuredeploy.log.$$ 2>&1
+  sudo sh -c "mkfs -t ext4 /dev/sdd" >> /tmp/azuredeploy.log.$$ 2>&1
+  sudo sh -c "mkdir /data" >> /tmp/azuredeploy.log.$$ 2>&1
+  sudo sh -c "mkdir /data2" >> /tmp/azuredeploy.log.$$ 2>&1
+  sudo sh -c "mount /data2" >> /tmp/azuredeploy.log.$$ 2>&1
+  echo "/dev/sdc /data ext4  defaults,discard 0 0" | sudo tee -a /etc/fstab >> /tmp/azuredeploy.log.$$ 2>&1
+fi
 sudo sh -c "mount /data" >> /tmp/azuredeploy.log.$$ 2>&1
 sudo sh -c "chown -R $ADMIN_USERNAME /data" >> /tmp/azuredeploy.log.$$ 2>&1
 sudo sh -c "chgrp -R $ADMIN_USERNAME /data" >> /tmp/azuredeploy.log.$$ 2>&1
